CLI-837 Fold declarative-integrations-migration.ts's wrapper into post-update.ts
After the reconcile logic moved to core/framework/features/reconcile.ts, all
that was left in this file was a 4-line load/reconcile/save wrapper -- not
enough to justify its own module and test file. Define
migrateDeclarativeIntegrations directly in post-update.ts instead, matching
its actual weight, and fold its two tests into post-update.test.ts.
